--- .env.prod.db
@@ -1,3 +1,0 @@
-POSTGRES_USER=misoadmin
-POSTGRES_PASSWORD=miso1234
-POSTGRES_DB=cloudtask
