--- .env.save
@@ -1,16 +1,0 @@
-localhostFLASK_APP="app.py"
-FLASK_DEBUG="1"
-HOSTNAME="0.0.0.0"
-DATABASE_URL="postgresql://misoadmin:miso1234@localhost:5432/cloudtask"
-UPLOAD_FOLDER ="/home/leslysharyn/audios/original"
-CONVERTED_FOLDER ="/home/leslysharyn/audios/converted"
-CELERY_BROKER_URL="redis://localhost:6379/0"
-SMTP_SERVER="smtp.mailtrap.io"
-SMTP_PORT="2525"
-SMTP_USERNAME="43016d8c11fc9e"
-SMTP_PASSWORD="ace826b1322476"
-JWT_SECRET_KEY="frase-secreta"
-SQL_HOST=db
-SQL_PORT=5432
-DATABASE=postgres
-SEND_EMAIL="False"
