--- Dockerfile.prod
@@ -1,88 +1,0 @@
-###########
-# BUILDER #
-###########
-
-# pull official base image
-FROM python:3.10.7-slim-buster as builder
-
-# set work directory
-WORKDIR /usr/src/app
-
-# set environment variables
-ENV PYTHONDONTWRITEBYTECODE 1
-ENV PYTHONUNBUFFERED 1
-
-# install system dependencies
-RUN apt-get update && \
-    apt-get install -y --no-install-recommends gcc
-
-# # lint
-# RUN pip install --upgrade pip
-# RUN pip install flake8==5.0.4
-# COPY . /usr/src/app/
-# RUN flake8 --ignore=E501,F401 .
-
-# install python dependencies
-COPY ./requirements.txt .
-RUN pip wheel --no-cache-dir --no-deps --wheel-dir /usr/src/app/wheels -r requirements.txt
-
-
-#########
-# FINAL #
-#########
-
-# pull official base image
-FROM python:3.10.7-slim-buster
-
-# create directory for the app user
-# RUN mkdir -p /usr/home/app
-
-# create the app user
-RUN addgroup --system app && adduser --system --group app
-
-# create the appropriate directories
-ENV HOME=/home/app
-ENV APP_HOME=/home/app/
-RUN mkdir /home/app/audios
-RUN mkdir /home/app/audios/uploaded
-RUN mkdir /home/app/audios/processed
-RUN mkdir /home/app/logs
-RUN touch /home/app/logs/celery.log
-RUN touch /home/app/logs/celerybeat.log
-# RUN mkdir $APP_HOME
-WORKDIR $APP_HOME
-
-# install dependencies
-RUN apt-get update && apt-get install -y --no-install-recommends netcat && apt-get -y install curl && apt-get -y install ufw && apt-get install -y ffmpeg
-COPY --from=builder /usr/src/app/wheels /wheels
-COPY --from=builder /usr/src/app/requirements.txt .
-RUN pip install --upgrade pip
-RUN pip install --no-cache /wheels/*
-
-#Configuring Firewall
-# RUN su ufw allow 5000
-
-# copy entrypoint-prod.sh
-COPY ./entrypoint.prod.sh $APP_HOME
-COPY ./cloud-miso-8.json $APP_HOME
-# copy project
-COPY . $APP_HOME
-
-# chown all the files to the app user
-RUN chown -R app:app $APP_HOME
-
-COPY ./celery/worker/start.sh /start-celeryworker
-RUN sed -i 's/\r$//g' /start-celeryworker
-RUN chmod +x /start-celeryworker
-
-COPY ./celery/beat/start.sh /start-celerybeat
-RUN sed -i 's/\r$//g' /start-celerybeat
-RUN chmod +x /start-celerybeat
-
-# RUN sed -i 's/\r$//g' /home/app/
-
-# change to the app user
-# USER app
-
-# run entrypoint.prod.sh
-ENTRYPOINT ["/home/app/entrypoint.prod.sh"]
