--- SuscriberService/modelo/__init__.py
@@ -1,1 +1,0 @@
-from .modelo import *
