--- celery/beat/start.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -o errexit
-set -o nounset
-
-rm -f './celerybeat.pid'
-cd ./tasks
-celery -A tasks beat -l info --logfile=../logs/celerybeat.log
