--- celery/worker/start.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-set -o errexit
-set -o nounset
-
-cd ./tasks
-celery -A tasks worker -l info --logfile=../logs/celery.log
