--- docker-compose.prod.yml
@@ -1,62 +1,0 @@
-version: '3.8'
-
-services:
-  web:
-    build:
-      context: ./
-      dockerfile: Dockerfile.prod
-    command: gunicorn --bind 0.0.0.0:5000 wsgi:app
-    expose:
-      - 5000
-    volumes:
-      - home_data:/home/app
-    env_file:
-      - ./.env.prod
-    depends_on:
-      - db
-    cpus: 2
-    mem_limit: '2GB'
-  db:
-    image: postgres:13-alpine
-    volumes:
-      - postgres_data_prod:/var/lib/postgresql/data/
-    env_file:
-      - ./.env.prod.db
-  nginx:
-    build: ./nginx
-    ports:
-      - 1337:80
-    depends_on:
-      - web
-  redis:
-    image: redis:6-alpine
-  
-  # celery_worker:
-  #   build:
-  #     context: .
-  #     dockerfile: Dockerfile.prod
-  #   image: flask_celery_worker
-  #   command: /start-celeryworker
-  #   env_file:
-  #     - ./.env.prod
-  #   depends_on:
-  #     - redis
-  #     - db
-  #   cpus: 2
-  #   mem_limit: '2GB'
-
-  celery_beat:
-    build:
-      context: .
-      dockerfile: Dockerfile.prod
-    image: flask_celery_beat
-    command: python ./tasks/tasks.py >> logs/file_logs.txt
-    env_file:
-      - ./.env.suscriber.prod
-    volumes:
-      - home_data:/home/app
-
-
-volumes:
-  postgres_data_prod:
-  home_data:
