--- docker-compose.yml
@@ -1,23 +1,0 @@
-version: '3.8'
-
-services:
-  web:
-    build: ./
-    command: python app.py run -h 0.0.0.0
-    volumes:
-      - ./:/usr/src/app/
-    ports:
-      - 5000:5000
-    env_file:
-      - ./.env
-  db:
-    image: postgres:13-alpine
-    volumes:
-      - postgres_data:/var/lib/postgresql/data/
-    environment:
-      - POSTGRES_USER=misoadmin
-      - POSTGRES_PASSWORD=miso1234
-      - POSTGRES_DB=cloudtask
-
-volumes:
-  postgres_data:
