--- nginx/Dockerfile
@@ -1,4 +1,0 @@
-FROM nginx:1.23-alpine
-
-RUN rm /etc/nginx/conf.d/default.conf
-COPY nginx.conf /etc/nginx/conf.d
