--- nginx/nginx.conf
@@ -1,16 +1,0 @@
-upstream cloud_task {
-    server web:5000;
-}
-
-server {
-
-    listen 80;
-    client_max_body_size 64M;
-    location / {
-        proxy_pass http://cloud_task;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        proxy_set_header Host $host;
-        proxy_redirect off;
-    }
-
-}
